Split PlayerActivity
PlayerActivityBase manages the controls -- things that came from the template.
PlayerActivity contains only the player-related code.
